requirements.txt
# Usage: pip install -r requirements.txt

# fedyolov7-iov requirements ---------------------
cryptography>=36.0.1
matplotlib>=3.5.1
mpi4py>=3.1.3
numpy>=1.24.2
opencv-python>=4.6.0
pandas>=1.4.0
Pillow>=9.0.1
PyYAML>=6.0
requests>=2.27.1
scipy>=1.8.0
seaborn>=0.12.2
torch>=1.13.1
torchvision>=0.14.1
tqdm>=4.65.0

# Other YOLOv7 requirements --------------------
## Base ----------------------------------------
protobuf<4.21.3

## Logging -------------------------------------
tensorboard>=2.12.0
# wandb

## Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

## Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
